refactor: 관리자 예약 추가 API End Point 수정으로 인한 리팩토링
