+++ README.md
@@ -336,7 +336,7 @@ id : long
 - [x] 사용자 예약 생성 기능 변경
   - 사용자가 예약 생성 시 로그인한 사용자 정보를 활용하도록 리팩터링
   - reservation.html, user-reservation.js 파일의 TODO 주석을 참고
-- [ ] 예약 생성 기능 변경 - 관리자
+- [x] 예약 생성 기능 변경 - 관리자
   - 관리자가 예약 생성 시 유저를 조회하여 선택 후 예약을 생성하도록 리팩터링
   - admin/reservation-new.html 파일에서 안내된 4단계 관련 주석에 따라 js 파일을 변경
     - AS-IS: /js/reservation-new.js
